+++ src/main/java/homeTry/exerciseList/exception/badRequestException/ExerciseAlreadyStartedException.java
@@ -1,6 +1,7 @@
-package homeTry.exerciseList.exception;
+package homeTry.exerciseList.exception.badRequestException;
 
 import homeTry.exception.BadRequestException;
+import homeTry.exerciseList.exception.ExerciseErrorType;
 
 public class ExerciseAlreadyStartedException extends BadRequestException {
