+++ src/main/java/homeTry/exerciseList/exception/badRequestException/ExerciseNameTooLongException.java
@@ -1,6 +1,7 @@
-package homeTry.exerciseList.exception;
+package homeTry.exerciseList.exception.badRequestException;
 
 import homeTry.exception.BadRequestException;
+import homeTry.exerciseList.exception.ExerciseErrorType;
 
 public class ExerciseNameTooLongException extends BadRequestException {
